Redesign reconciliation of padding requirements
It is now done generically through an interface in accel, rather than
being special-cased in FlaggerDevice. Additionally, all operations have
been partitioned into a template (which contains the program code) and
an instance (which has fixed sizes, and is bound to specific buffers).
This removes the need for handling on-the-fly reallocation of temporary
buffers.

This has necessitated having separate stride parameters for different
inputs in several kernels, to avoid complicated padding requirements
that spanned multiple buffers.

The transposition kernel has been moved into a separate file.

While it looks like this has added a lot of code to device.py, the
additional are almost entirely documentation. The amount of actual code
is broadly similar.

While the current test suite passes, more tests are needed to
specifically test the new padding code, and possibly to enable forced
extra padding.