Redesign slots to use Dimension
This allows multiple slots to share the same dimension, which then
requires only a single stride to be passed into kernels. It also allows
dimensions to be specified as exact, so that complex numbers can be
packed into a size-2 dimension with a guarantee of no padding between
the complex values.