Redesign testing of autotuning
@device_test now suppresses autotuning, and returns a fixed value that
must be provided by the class author (as an argument to @autotuner). A
new @force_autotune decorator can re-enable it, but disables the cache
so that it will run regardless.

Finally, the cache itself is tested in test_autotune, but with an
in-memory database. Thus, unit tests never touch an on-disk database.

This also fixes a bug where the autotuner for
BackgroundMedianFilterDeviceTemplate with amplitudes=True didn't work.