Change @device_test decorator to pass context and command_queue to the test function
This fixes an issue where merely importing test_accel would cause
initialisation (including printing the device name) to occur, even if
the import was being doing by a documentation or static analysis tool.

It also means that it is no longer possible to accidentally omit the
@device_test decorator.

I initially tried leaving the variables as global, but only initialising
them when the first test is run (and setting them back to None after
each test). Unfortunately, the way Python interprets
"from test_accel import test_context" means that the current module gets
a *copy* of the variable, which is unaffected by these runtime
manipulations.