Generalize the exact=True case for accel.Dimension
Previously exact=True could only be specified when no padding was
present. Now, it is possible to specify it when padding is present, and
it specifies that the padding must be precisely that for this dimension,
and no further padding can be added. This is necessary to support CUFFT
in katsdpimager, where the padding is baked into the Plan object.