Make the rfi flagging take account of previously defined flags when determining new ones.
This commit modifies the sumthreshold flagging and spline backgrounding so that it optionally
ignores predefined flags input by the user. Other flagging algorithms will be updated later.
This commit also splits the flag report function into a function that determines the flags
and writes them to an h5 file and a separate function that reads flags from the output h5 file
(or the katdal object if the output file is not available) and produces the pdf report.
This should allow for RTS scripts to read in these determined flag tables if the flags from
ingest are indufficient.