Support shuffle reduce in more cases
Sizes that are a multiple of the warp size but not a power of 2 are now
supported.

While it should in theory be possible to support other sizes greater
than the warp size (but not a multiple of it), this only works when
the participating threads are aligned to the warps, so is not supported
for now.