Allow autotune configuration to be explicitly skipped
Previously they could be skipped by throwing an exception. However, when
digging through logs these fake exceptions were cluttering the logs,
making it hard to spot the serious/unintended exceptions. Callers to
autotune can now return None to skip a configuration, and doing so is
silent in the log so that one can see the interesting stuff more easily.