Allow templates to be shared between threads
Enqueuing a kernel is not thread-safe, due to the way it is implemented
in OpenCL (sets arguments on the object then invokes it). The kernel
objects are now moved from the templates to the operations, so that a
single template can be instantiated and then the instances can be used
in separate threads. The program is still kept in the template, since
that is a heavyweight object.

Closes SR-131.