Add option to backgrounder to ignore already-flagged data
The median filter only considers non-flagged values when computing the
median. This is to pave the way for feeding operator channel
masks into the ingest flagging process.

At the moment the channel flags are not folded into the output. It would
add a lot more complexity to handle it in each of the thresholders,
whereas ingest already has code to do it later in the pipeline.