Change exceptions handling in resource context manager
The previous behaviour was to cancel the future that the next user
waited on, so that it wouldn't block forever. In some cases that would
lead to a CancelledError propagating outwards before the original
exception, leading to hard-to-debug errors.

The new behaviour sets the original exception as the exception on the
future. This preserves the behaviour that the next waiter gets woken up
and aborted, but now the logged exceptions are friendlier.