Make katsdpsigproc work on Python 3
This is mostly just a case of applying the changes from
python-modernize. There are a few places where the change is a little
trickier, because Python 3 changes sys.exc_info() so that it only
returns values while the exception handler is active, rather than always
giving access to the last exception.