Allow twodflag caller to supply own pool
Instead of taking a number of cores (which are used to create a
multiprocessing pool), SumthresholdFlagger now takes a pool. This allows
the caller to integrate the flagger with its own pool, which could be of
any type that implements the asynchronous interface.

It's also possible to specify no pool (the default) if the caller is
doing its own chunking of the data.

The method for operating on a single baseline is also renamed from the
cryptic detect_spikes_sumthreshold to the more natural
get_baseline_flags.