Redo handling of frequency chunks in 2D flagger
Previously the band was split into sub-bands at the top level of the
flagger, and each sub-band flagged separately. This required substantial
padding to avoid edge effects, and in fact the padding used wasn't
sufficient to fully hide the edges. The results were also sensitive to
the amount of padding used, because the padding values contributed to
the noise estimates.

The new approach pushes the subbands into getbackground_2d and
_sumthreshold, where they are explicitly used just for noise estimates.
_sumthreshold still uses the chunking + padding approach because the
clamping step really does need to be executed multiple times at the
boundaries (once for each applicable threshold).

This will slightly affect the results, because noise estimates are more
local than previously. The number of subbands could be reduced to
compensate.