Rewrite of twodflag for performance
95% of it now runs inside of numba (with nogil=True), so it scales well
in multithreading. It would still be possible to do slightly more inside
numba (the _get_flags function), but it will be only make a very small
difference.

There may be some small changes in results due to steps being reordered
(in particular, the detection of frequencies/times which are almost all
flagged), but the results should be more-or-less the same.

There are still several opportunities for optimisation, particularly
where memory access is along columns (which will be inefficient).