Add support for using asyncio instead of trollius
The asyncio versions are in katsdpsigproc.asyncio subpackage. Some
trickery is used in setup.py to prevent these being installing in Python
< 3.4, even though they are always present in the sdist.

The generation of asyncio versions currently requires a manual step to
run update-asyncio.sh. The advantage is that it avoids a setup_requires
dependency, plus I'm not too sure how to do such automatic conversions
in setup.py and have it work correctly for all cases.