Support latest version of pycuda
Since `async` is a keyword in Python 3.7, pycuda has renamed a parameter
to `async_`. It tries to have backwards compatibility but screwed it up.

To support both old and new pycuda, use the get_async and set_async
methods, which are just thin wrappers around what we were doing anyway.