Rip out Python 2 and Trollius support
This simplies things mainly because there is no longer a need to keep
separate copies of resource.py for trollius and for asyncio (with a
script to recreate one from the other).

katsdpsigproc.resource is now the asyncio version instead, and
katsdpsigproc.asyncio.resource is a deprecated alias to it.

The use of six is also removed.