Change Operation.buffer to raise if given an alias slot.
Before the previous round of type annotation work, it would have
raised AttributeError (because AliasIOSlot doesn't have a `buffer`
attribute), and I changed it to return None instead. However, None
should only be used to indicate that it's a valid slot for a buffer but
none has been bound yet. This shouldn't break compatibility because it's
only very briefly returned None, so no code should be depending on it.