Use BLOCKING_SYNC for events
This makes waiting for events interrupt-driven rather than spinning on
the CPU. In most cases users of katsdpsigproc have other work to do (or
other processes making use of CPU) so this is friendlier. It will
increase latency fractionally, but I don't think it's typically being
used with tiny kernels so I'm not worried.