zscale: use RMS not np.std to get rejection threshold
When the data is *very* linear (which was happening in a unit test), the
residuals weren't centred around zero due to rounding error (normally
I think they would be because that minimises the least-squares), so
np.std was underestimating the RMS of the residuals, causing everything
to get rejected.

For some reason this was only happening in particular environments -
maybe due to instruction set differences e.g. AVX2.