Make zscale return nan for vmin and vmax when the input sample array is empty.
This is most likely to occur when the input image is completely non-finite.
Returning nan's should work here since the vmin, vmax parameters of
pyplot.imshow can be nans, in which case the scaling is just set to some default
and allows completely blanked FITS images to be displayed when scaled
by the zscale algorithm.