Suppress asyncio warning about non-retrieved exception
When an exception is thrown inside a ResourceAllocation context
management block, it is copied into a future, but also allowed to
propagate. Since the copy in the future is just a duplicate, we don't
want to complain if it is never consumed. This was causing some spurious
timing-dependent test failures (depending on when it got
garbage-collected).