Replace type comments with type annotations
This was mostly done automatically by the com2ann package, but some
annotations were modified to change

x: T = None

to

x: T

This was needed where T is not an optional type. In these cases the None
assignment was just a placeholder in an abstract base class, with the
concrete class filling in an instance of T during the constructor.